bugfix:优化dashboard查询速率

+++ arius-admin/arius-admin-common/src/main/java/com/didichuxing/datachannel/arius/admin/common/bean/entity/metrics/linechart/MetricsContent.java
@@ -37,6 +37,6 @@ public class MetricsContent implements Serializable {
     /**
      * 集群索引数量
      */
-    private Long                   indexCount;
+    private Double                   indexCount;
 
 }
